pyproject.toml
[project]
name = "s3-image-converter"
version = "0.1.0"
description = ""
authors = [
    {name = "Suyash Shrestha",email = "sthasuyash11@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3 (>=1.39.7,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-mock (>=3.14.1,<4.0.0)",
    "pytest-cov (>=6.2.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
